Do not consult base_class for queries
We use acts_as_nested_set in a subclass, which inherits from a model.
With that setup, Rails assumes STI, but it is not the case for us.
We always want to use the subclass and not the superclass, e.g. when
asking for the root of a node.

Therefore we "fix" this gem by "breaking" the support for STI. Ugly, but
it works for now until the code on our end is refactored.